+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Full changelog for the Transmute Coffee website.
 
+## v1.62.0 - New releases - 2019-05-31
+
+* New release, Costs to Expect Website, v1.03.1.
+* New release, Costs to Expect Web app, v1.04.3.
+* New release, Costs to Expect API, v1.14.2
+
 ## v1.61.0 - New release - 2019-05-10
 
 * New release, Costs to Expect Website, v1.02.0.